seperate email logic into own service
