+++ backend/src/routes/auth-route.ts
@@ -2,10 +2,9 @@ import express, { Request, Response } from "express";
 import { User } from "../model/schemas/userSchema";
 import { UserRepository } from "../repository/UserRepository";
 import { isValidUsername } from "../utils/utils";
+import { sendWelcomeEmail } from "../services/EmailService";
 
 const router = express.Router();
-
-const nodeMailer = require("nodemailer");
 
 const userRepository = new UserRepository();
 
@@ -88,33 +87,8 @@ router.post("/signup", async (req: Request, res: Response) => {
       req.session.username = newUser.username;
       req.session.save();
 
-      // send email
-      var transporter = nodeMailer.createTransport({
-        service: "gmail",
-        auth: {
-          user: "kodygramme@gmail.com",
-          pass: "buaknommahtbfzpu",
-        },
-      });
+      await sendWelcomeEmail("");
 
-      // Update the text property to include an anchor tag with the YouTube link
-      var mailOptions = {
-        from: "kodygramme@gmail.com",
-        to: "zhiani2000@gmail.com", // shramko.georgiy@gmail.com BRUHHHH
-        subject: "You have been invited to collaborate on CodeGram",
-        html: "<p>@shaygeko has invited you to collaborate on CodeGram, <a href='https://www.youtube.com/watch?v=j5a0jTc9S10&list=PL3KnTfyhrIlcudeMemKd6rZFGDWyK23vx&index=11&ab_channel=YourUncleMoe'>click to join</a></p>",
-      };
-
-      transporter.sendMail(
-        mailOptions,
-        function (error: any, info: { response: string }) {
-          if (error) {
-            console.log(error);
-          } else {
-            console.log("Email sent: " + info.response);
-          }
-        }
-      );
 
       res.status(200).send("Registered");
     }
